Rename modal_iface -> modal_interface, move log_utils/guards to modal_proxy, restore FOLLOWUP
Three issues fixed:

1. Rename modal_iface -> modal_interface everywhere (55 occurrences
   across 4 files). The abbreviation was unclear and saved negligible
   characters.

2. Move log_utils.py and register_guards.py to modal_proxy so the
   functionality is available for future testing/remote implementations.
   The mng_modal files become thin re-exports. Tests import directly
   from modal_proxy. Added mng and resource-guards as modal_proxy deps.

3. Restore lost FOLLOWUP comments about Modal's tag-based sandbox
   filtering bug in _lookup_sandbox_by_host_id_once and
   _lookup_sandbox_by_name_once. These document that we're working
   around a Modal bug and should use tags={} filtering once fixed.
