+++ libs/mng_modal/imbue/mng_modal/log_utils_test.py
@@ -1,9 +1,9 @@
 from io import StringIO
 
-from imbue.mng_modal.log_utils import ModalLoguruWriter
-from imbue.mng_modal.log_utils import _create_modal_loguru_writer
-from imbue.mng_modal.log_utils import _create_multi_writer
-from imbue.mng_modal.log_utils import enable_modal_output_capture
+from imbue.modal_proxy.log_utils import ModalLoguruWriter
+from imbue.modal_proxy.log_utils import _create_modal_loguru_writer
+from imbue.modal_proxy.log_utils import _create_multi_writer
+from imbue.modal_proxy.log_utils import enable_modal_output_capture
 
 
 def test_multi_writer_writes_to_all_files() -> None:
