+++ libs/modal_proxy/pyproject.toml
@@ -6,7 +6,9 @@ readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
     "imbue-common",
+    "mng==0.1.7",
     "modal==1.3.1",
+    "resource-guards",
 ]
 
 [build-system]
@@ -18,6 +20,8 @@ packages = ["imbue"]
 
 [tool.uv.sources]
 imbue-common = { workspace = true }
+mng = { workspace = true }
+resource-guards = { workspace = true }
 
 [tool.pytest.ini_options]
 timeout_func_only = true
